requirements.txt
streamlit
pandas
matplotlib
plotly>=5.0.0
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client